third_party/WebKit/LayoutTests/fast/css/pseudo-out-of-range-expected.txt

Tests that we find controls if they have a range limitation and are out-of-range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

==> :out-of-range doesn't match to an INPUT without a value:
PASS document.querySelector("input[type=number]:out-of-range") is null


==> :out-of-range should match to an INPUT with an out-of-range value:
PASS document.querySelector("input[type=number]:out-of-range").id is "number1"
PASS document.querySelectorAll(":out-of-range").length is 1


==> When the value becomes in-range dynamically, we do not find the control anymore:
PASS document.querySelector("input[type=number]:out-of-range") is null
PASS document.querySelectorAll(":out-of-range").length is 0
PASS successfullyParsed is true

TEST COMPLETE
